Add long type cast for tweet id dict long key

+++ controller/tweet_db_utils.py
@@ -1,2 +1,2 @@
 def db_tweets_to_tweets_dict(tweets_list):
-    return {tweet['_id']: tweet for tweet in tweets_list}
+    return {long(tweet['_id']): tweet for tweet in tweets_list}
